fix: correctly detect screenshot name
Previously, we would automatically set the name of the screenshot to the most recently played game. This is incorrect for cases where we are not currently _in_ a game, such as in the main ui.

To fix this, we now detect minarch.elf - as well as a slew of popular optional emulators - when checking if we should use the game name or a generic name. While this is a bit more annoying to upkeep - especially as folks port other custom emulators - this is a bit better as now we don't errantly mark a screenshot of the UI as a game.

Ideally we also detect if another custom pak is launched and use the pak's name as the screenshot name, but this is better than nothing.