feat: implement messages when launching the screenshot monitor
Instead of just exiting immediately, show a message stating that it is being launched, check if it is running, and then exit as appropriate.

Additionally, update process tracking to handle the case where we are launching a script - which doesn't respond as nicely to pgrep - vs a binary.

Finally, rename a variable to make it make more sense.